Delegation-Backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-environ==0.11.2
python-decouple==3.8
channels==4.1.0
channels-redis==4.2.0
redis==5.0.1
psycopg2-binary==2.9.9
django-extensions==3.2.3
Pillow==10.3.0
whitenoise==6.6.0
gunicorn==21.2.0
